Move promiseAnimationFrame to separate file
